Conflicts:
	nbproject/project.properties

+++ web/empleado/cuenta/list.js
@@ -1,31 +1,41 @@
-function CuentaListController($scope, $http) {
-        
-    $scope.findAll = function() {
+app.controller("CuentaListController", function ($scope, $http) {
+
+    $scope.findAll = function () {
         $http({
             method: "GET",
-            url: contextPath + "/api/cuenta/"
-        }).success(function(data) {
+            url: contextPath + "/api/cuenta"
+        }).success(function (data) {
             $scope.cuentas = data;
-        }).error(function(data, status) {
+        }).error(function (data, status) {
             alert("Fatal error: " + status);
         });
     };
-    
+
     $scope.findAll();
 
-    $scope.deleteCuenta = function(id) {
-        $http({
-            method: "DELETE",
-            url: contextPath + "/api/cuenta/" + id
-        }).success(function() {
+    $scope.deleteCuenta = function (id) {
+        ok = confirm("¿ Está seguro que quiere borrar la cuenta de ID: " + id + " ?");
+
+        if (ok) {
+            $http({
+                method: "DELETE",
+                url: contextPath + "/api/cuenta/" + id
+            }).success(function () {
+                $scope.findAll();
+            }).error(function (data, status) {
+                alert("Fatal error: " + status);
+            });
+        } else {
             $scope.findAll();
-        }).error(function(data, status) {
-            alert("Fatal error: " + status);
-        });
-    };
-    
-    $scope.edit = function(id) {
-        location.replace('#/cuenta/edit/' + id);
+        } 
     };
 
-}
+    $scope.crearCuenta = function () {
+        location.replace("#/cuenta/insert");
+    };
+
+    $scope.editCuenta = function (id) {
+        location.replace("#/cuenta/edit/" + id);
+    };
+
+});
